refactor(FolderDataGrid): rename Table folder to DataGrid
