+++ packages/admin/cms-admin/src/dam/DataGrid/thumbnail/AudioThumbnail.tsx
