+++ packages/admin/cms-admin/src/dam/DataGrid/thumbnail/DamThumbnail.gql.ts
