+++ packages/admin/cms-admin/src/dam/DataGrid/thumbnail/DamThumbnail.tsx
