+++ packages/admin/cms-admin/src/dam/DataGrid/thumbnail/VideoThumbnail.tsx
