+++ packages/admin/cms-admin/src/dam/DataGrid/useDamSearchHighlighting.ts
