+++ packages/admin/cms-admin/src/dam/FileForm/EditFile.tsx
@@ -34,7 +34,7 @@ import {
     GQLUpdateFileMutation,
     GQLUpdateFileMutationVariables,
 } from "../../graphql.generated";
-import { usePersistedDamLocation } from "../Table/RedirectToPersistedDamLocation";
+import { usePersistedDamLocation } from "../DataGrid/RedirectToPersistedDamLocation";
 import Duplicates from "./Duplicates";
 import { damFileDetailQuery, updateDamFileMutation } from "./EditFile.gql";
 import { FilePreview } from "./FilePreview";
